Stress test utility for laboratory measurements
This commit adds a stress test utility called revpi-stress-tester, which is
intended to use during laboratory measurement with new RevPi hardware.

With programm parameter the user can spawn stress test worker for
the following parts of the hardware:

* CPU (via stress-ng)
* Network interfaces (via iperf3, RX and TX)

Furthermore the following metrics are gathered during test run:

* CPU clock speed
* CPU temperature
* Has the CPU been throttled

In Addition to that the rt latency can be monitored with cyclictest too.

Signed-off-by: Nicolai Buchwitz <n.buchwitz@kunbus.com>