fix(picontrol-usage): Make Ctrl+C work.
The use and forwarding of multiprocessing.Event created a deadlock
situation when receiving signal SIGINT. multiprocessing.Event uses
locks: Internally uses multiprocessing.Event locks and semaphores for
synchronization between processes. If the signal handler is called
during a lock operation and tries to access the event itself, a deadlock
occurs.

The solution is now to see the multiprocess as a own process, which
itself manages the signal handlers and events. In addition, terminate
and kill were built into the main process in order to be able to
terminate the processes internally in the future.

Signed-off-by: Sven Sager <s.sager@kunbus.com>